Get rid of `model.assembleScales()`
so we have less methods in models + have all assembleScale tests in one place

(We couldn't do this earlier due to circular dependencies problem.)
