+++ test/compile/layer.test.ts
@@ -57,35 +57,6 @@ describe('Layer', function() {
     });
   });
 
-  describe('assembleScales', () => {
-    it('includes all scales from children, both shared and independent', () => {
-      const model = parseLayerModel({
-        layer: [{
-          mark: 'point',
-          encoding: {
-            x: {field: 'a', type: 'quantitative'},
-            y: {field: 'c', type: 'quantitative'}
-          }
-        },{
-          mark: 'point',
-          encoding: {
-            x: {field: 'b', type: 'quantitative'},
-            y: {field: 'c', type: 'quantitative'}
-          }
-        }],
-        resolve: {
-          scale: {
-            x: 'independent'
-          }
-        }
-      });
-
-      model.parseScale();
-      const scales = model.assembleScales();
-      assert.equal(scales.length, 3); // 2 x, 1 y
-    });
-  });
-
   describe('dual axis chart', () => {
     const model = parseLayerModel({
       layer: [{
