+++ test/compile/repeat.test.ts
@@ -7,26 +7,6 @@ import {keys} from '../../src/util';
 import {parseRepeatModel} from '../util';
 
 describe('Repeat', function() {
-  describe('assembleScales', () => {
-    it('includes all scales', () => {
-      const model = parseRepeatModel({
-        repeat: {
-          row: ['Acceleration', 'Horsepower']
-        },
-        spec: {
-          mark: 'point',
-          encoding: {
-            x: {field: {repeat: 'row'}, type: 'quantitative'}
-          }
-        }
-      });
-
-      model.parseScale();
-      const scales = model.assembleScales();
-      assert.equal(scales.length, 2);
-    });
-  });
-
   describe('resolveRepeat', () => {
     it('should resolve repeated fields', () => {
       const resolved = replaceRepeaterInEncoding({
