+++ test/compile/scale/assemble.test.ts
@@ -1,8 +1,100 @@
 import {assert} from 'chai';
-import {assembleScaleRange} from '../../../src/compile/scale/assemble';
-import {parseUnitModel, parseUnitModelWithScale} from '../../util';
+import {assembleScaleRange, assembleScales} from '../../../src/compile/scale/assemble';
+import {parseConcatModel, parseFacetModelWithScale, parseLayerModel, parseRepeatModel, parseUnitModel, parseUnitModelWithScale} from '../../util';
 
 describe('compile/scale/assemble', () => {
+  describe('assembleScales', () => {
+    it('includes all scales for concat', () => {
+      const model = parseConcatModel({
+        vconcat: [{
+          mark: 'point',
+          encoding: {
+            x: {field: 'a', type: 'ordinal'}
+          }
+        },{
+          mark: 'bar',
+          encoding: {
+            x: {field: 'b', type: 'ordinal'},
+            y: {field: 'c', type: 'quantitative'}
+          }
+        }]
+      });
+
+      model.parseScale();
+      const scales = assembleScales(model);
+      assert.equal(scales.length, 3);
+    });
+
+
+    it('includes all scales from children for layer, both shared and independent', () => {
+      const model = parseLayerModel({
+        layer: [{
+          mark: 'point',
+          encoding: {
+            x: {field: 'a', type: 'quantitative'},
+            y: {field: 'c', type: 'quantitative'}
+          }
+        },{
+          mark: 'point',
+          encoding: {
+            x: {field: 'b', type: 'quantitative'},
+            y: {field: 'c', type: 'quantitative'}
+          }
+        }],
+        resolve: {
+          scale: {
+            x: 'independent'
+          }
+        }
+      });
+
+      model.parseScale();
+      const scales = assembleScales(model);
+      assert.equal(scales.length, 3); // 2 x, 1 y
+    });
+
+    it('includes all scales for repeat', () => {
+      const model = parseRepeatModel({
+        repeat: {
+          row: ['Acceleration', 'Horsepower']
+        },
+        spec: {
+          mark: 'point',
+          encoding: {
+            x: {field: {repeat: 'row'}, type: 'quantitative'}
+          }
+        }
+      });
+
+      model.parseScale();
+      const scales = assembleScales(model);
+      assert.equal(scales.length, 2);
+    });
+
+    it('includes shared scales, but not independent scales (as they are nested) for facet.', () => {
+      const model = parseFacetModelWithScale
+      ({
+        facet: {
+          column: {field: 'a', type: 'quantitative', format: 'd'}
+        },
+        spec: {
+          mark: 'point',
+          encoding: {
+            x: {field: 'b', type: 'quantitative'},
+            y: {field: 'c', type: 'quantitative'}
+          }
+        },
+        resolve: {
+          scale: {x: 'independent'}
+        }
+      });
+
+      const scales = assembleScales(model);
+      assert.equal(scales.length, 1);
+      assert.equal(scales[0].name, 'y');
+    });
+  });
+
   describe('assembleScaleRange', () => {
     it('replaces a range step constant with a signal', () => {
       const model = parseUnitModel({
